backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==0.61.1
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
frozendict==2.4.6
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
joblib==1.4.2
lxml==5.3.0
multitasking==0.0.11
numpy==2.0.2
oauthlib==3.2.2
openai==1.55.0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pillow==11.0.0
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.4.0
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webencodings==0.5.1
xmltodict==0.14.2
yfinance==0.2.50
